+++ lima_common/src/common/misc/fsaStringsPool.cpp
@@ -48,7 +48,7 @@ class FsaStringsPoolPrivate
   FsaStringsPoolPrivate(const FsaStringsPool& /*unused p*/);
 
   void clear();
-  void clear(const uint32_t pos);
+  void clear(const uint64_t pos);
 
   Common::AbstractAccessByString* m_mainKeys;
   StringsPoolIndex m_mainKeySize;
